// src/config/env.d.ts
export declare const env: {
    PORT: string;
    NODE_ENV: "development" | "production" | "test";
    DATABASE_URL: string;
    JWT_SECRET: string;
    JWT_EXPIRES_IN: string;
    CLOUDINARY_CLOUD_NAME?: string | undefined;
    CLOUDINARY_API_KEY?: string | undefined;
    CLOUDINARY_API_SECRET?: string | undefined;
    FRONTEND_URL?: string | undefined;
    MAIL_FROM?: string | undefined;
    MAIL_PROVIDER_API_KEY?: string | undefined;
    APP_URL?: string | undefined;
};
//# sourceMappingURL=env.d.ts.map